Move replaceWordInline to core API
