+++ src/muya/lib/index.js
@@ -398,44 +398,45 @@ class Muya {
    * @param {boolean} setCursor Shoud we update the editor cursor?
    */
   replaceWordInline (line, wordCursor, replacement, setCursor = false) {
-    const { start: lineStart, end: lineEnd } = line
-    const { start: wordStart, end: wordEnd } = wordCursor
+    this.contentState.replaceWordInline(line, wordCursor, replacement, setCursor)
+    // const { start: lineStart, end: lineEnd } = line
+    // const { start: wordStart, end: wordEnd } = wordCursor
 
-    // Validate cursor ranges.
-    if (wordStart.key !== wordEnd.key) {
-      throw new Error('Expect a single line word cursor: "start.key" is not equal to "end.key".')
-    } else if (lineStart.key !== lineEnd.key) {
-      throw new Error('Expect a single line line cursor: "start.key" is not equal to "end.key".')
-    } else if (wordStart.offset > wordEnd.offset) {
-      throw new Error(`Invalid word cursor offset: ${wordStart.offset} should be less ${wordEnd.offset}.`)
-    } else if (lineStart.key !== wordEnd.key) {
-      throw new Error(`Cursor mismatch: Expect the same line but got ${lineStart.key} and ${wordEnd.key}.`)
-    } else if (lineStart.block.text.length < wordEnd.offset) {
-      throw new Error('Invalid cursor: Replacement length is larger than line length.')
-    }
+    // // Validate cursor ranges.
+    // if (wordStart.key !== wordEnd.key) {
+    //   throw new Error('Expect a single line word cursor: "start.key" is not equal to "end.key".')
+    // } else if (lineStart.key !== lineEnd.key) {
+    //   throw new Error('Expect a single line line cursor: "start.key" is not equal to "end.key".')
+    // } else if (wordStart.offset > wordEnd.offset) {
+    //   throw new Error(`Invalid word cursor offset: ${wordStart.offset} should be less ${wordEnd.offset}.`)
+    // } else if (lineStart.key !== wordEnd.key) {
+    //   throw new Error(`Cursor mismatch: Expect the same line but got ${lineStart.key} and ${wordEnd.key}.`)
+    // } else if (lineStart.block.text.length < wordEnd.offset) {
+    //   throw new Error('Invalid cursor: Replacement length is larger than line length.')
+    // }
 
-    const { block } = lineStart
-    const { offset: left } = wordStart
-    const { offset: right } = wordEnd
+    // const { block } = lineStart
+    // const { offset: left } = wordStart
+    // const { offset: right } = wordEnd
 
-    // Replace word range with replacement.
-    block.text = block.text.substr(0, left) + replacement + block.text.substr(right)
+    // // Replace word range with replacement.
+    // block.text = block.text.substr(0, left) + replacement + block.text.substr(right)
 
-    // Update cursor
-    if (setCursor) {
-      const cursor = Object.assign({}, wordStart, {
-        offset: left + replacement.length
-      })
-      line.start = cursor
-      line.end = cursor
-      this.contentState.cursor = {
-        start: cursor,
-        end: cursor
-      }
-    }
+    // // Update cursor
+    // if (setCursor) {
+    //   const cursor = Object.assign({}, wordStart, {
+    //     offset: left + replacement.length
+    //   })
+    //   line.start = cursor
+    //   line.end = cursor
+    //   this.contentState.cursor = {
+    //     start: cursor,
+    //     end: cursor
+    //   }
+    // }
 
-    this.contentState.partialRender()
-    this.dispatchChange()
+    // this.contentState.partialRender()
+    // this.dispatchChange()
   }
 
   destroy () {
